make modified by default + fix bug
This commit makes the default behavior of the render script to only
render the modified slides.

(Ideally we would hash the md files and check against a database
if it needs to be re-rendered, or use the git last commit time
instead, but we can't have nice things for now)

Also re-renders everything to make update times nice.